nuxt-property-decoratorを完全削除
