+++ client/components/CookieFooter.vue
@@ -7,7 +7,7 @@
                 v-on:accept="$store.commit('acceptCookie')"
             >
                 <div slot="message">
-                    {{ message }}
+                    Nuxt Adminを利用すると、Cookieの使用に同意したことになります。jiyuujin LAB.とそのパートナーは、アナリティクス、カスタマイズ、広告目的も含めて、世界中でサービスを提供しCookieを使用します。
                 </div>
             </cookie-law>
         </client-only>
@@ -15,15 +15,12 @@
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'nuxt-property-decorator'
+import Vue from 'vue'
 import CookieLaw from 'vue-cookie-law'
 
-@Component({
+export default Vue.extend({
     components: {
         CookieLaw
-    },
+    }
 })
-export default class CookieFooter extends Vue {
-    message: string = 'Nuxt Adminを利用すると、Cookieの使用に同意したことになります。jiyuujin LAB.とそのパートナーは、アナリティクス、カスタマイズ、広告目的も含めて、世界中でサービスを提供しCookieを使用します。'
-}
 </script>
