+++ package.json
@@ -30,7 +30,6 @@
     "graphql": "14.6.0",
     "graphql-tag": "2.10.3",
     "nuxt": "2.11.0",
-    "nuxt-property-decorator": "2.5.0",
     "ts-node": "8.6.2",
     "vue-chartjs": "3.5.0",
     "vue-cookie-law": "1.11.0",
